Refactor paging feature into its own file
